Bug 884: Add standard JogAmp build files.
The new make/build.xml successfully builds and tests the project. The
original NetBeans build files are still present, but won't work anymore
since the resources directory is now inside the make directory. The new
build files are based on those of JOAL. The new build is not yet minimized
and cleaned up, because this commit is meant to be the minimum way to get
things functional.